requirements.txt
gitdb==4.0.11
GitPython==3.1.43
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
rich==13.8.1
smmap==5.0.1
